Updates for Spring 2026

+++ assignment-spark-wordcount/src/main/scala/edu/neu/coe/csye7200/asstswc/WordCount_ScalaVanilla.scala
@@ -6,7 +6,7 @@ object WordCount_ScalaVanilla {
     //Use scala original methods to perform word count
     //Result format should similiar to Map(Hello -> 3, Hi -> 1, World -> 3)
     //hint: you may start with flatMap() then consider the steps to process data in Spark
-    def wordCount(list: List[String]): Map[String, Int] = {
+    def wordCount(list: List[String]): Map[String, Long] = {
 // TO BE IMPLEMENTED 
 ???
     }
